rename: 회원가입 스키마 파일 이동
